// types.ts
export interface GalleryItem {
  id: string;
  url: string;
  title: string;
  titleZh: string;
  phrase: string;
  phraseZh: string;
  date: string;
  location: string;
  locationZh: string;
  depth: number;
  parallaxFactor: number;
}

export interface ImageData {
  url: string;
  title: string;
  titleZh: string;
  phrase: string;
  phraseZh: string;
  date: string;
  location: string;
  locationZh: string;
}